Check persisted records not previous
A last minute change to find_each caused these
tests to start failing. This was because find_each
initializes a new copy of the labwares, regardless
of whether the association was previously populated.
Meanwhile the tests were looking at the original instance,
which is now stale.

This change fetches the labware from the database
and looks at those, rather than reverting the find behaviour.

I made this decision because:
1 - It will better handle the emptying of very large locations
without causing memory spikes.
2 - It is the persisted labware we care about, it would
be bad if we updated the in-memory version but didn't
persist it.
3 - As the code is actually used, the labware association isn't
loaded, so this requirement is unique to the tests.

It would be possible to do the following:
```
iteration_method = labwares.loaded? ? :each : :find_each
labwares.send(iteration_method) do
...
```

To ensure we update existing labwares if they are loaded,
or fetch them efficiently if they aren't.
However I felt this complexity was unecessary.